requirements.txt
fastapi
uvicorn[standard]
httpx
tiktoken
python-dotenv
pydantic
supabase